+++ src/main/java/org/bytedeco/javacpp/tools/Parser.java
@@ -102,8 +102,8 @@ public class Parser {
         return s.substring(s.lastIndexOf(' ') + 1);
     }
 
-    static String removeVariadic(String s) {
-        return s.endsWith("...") ?  s.substring(0, s.length()-3) + "[]" : s;
+    static String desugarVarargs(String s) {
+        return s.trim().endsWith("...") ? s.trim().substring(0, s.length() - 3) + "[]" : s;
     }
 
     static String upcastMethodName(String javaName) {
@@ -2523,7 +2523,7 @@ public class Parser {
                     extraDecl = new Declaration();
                     extraDecl.signature = signature;
                     extraDecl.declarator = dcl; // Used in group to recognize friends
-                    extraDecl.text = "public " + removeVariadic(dcl.type.javaName) + " " + dcl.javaName + "(" + argList + ") { "
+                    extraDecl.text = "public " + desugarVarargs(dcl.type.javaName) + " " + dcl.javaName + "(" + argList + ") { "
                               + (dcl.type.javaName.equals("void") ? "" : "return ") + dcl.javaName + "(" + staticArgList + "); }\n";
                 } else {
                     friendly = false;
@@ -2601,13 +2601,13 @@ public class Parser {
                         sb.append(removeAnnotations(param.type.javaName)).append(" ").append(param.javaName);
                     }
                     decl.text += accessModifier + " " + (staticMethod ? "static " : "")
-                              +  removeVariadic(removeAnnotations(type.javaName)) + " " + dcl.javaName + "(" + sb + ") { "
+                              +  desugarVarargs(removeAnnotations(type.javaName)) + " " + dcl.javaName + "(" + sb + ") { "
                               +  (type.javaName.equals("void") ? "" : "return ")
                               +  (context.upcast && !staticMethod ? upcastMethodName(context.javaName) + "()." : "")
                               +  "_" + dcl.javaName + (dcl.parameters.names == null ? "()" : dcl.parameters.names) + "; }\n";
                     dcl.javaName = "_" + dcl.javaName;
                 }
-                decl.text += modifiers2 + type.annotations + context.shorten(removeVariadic(type.javaName)) + " " + dcl.javaName + dcl.parameters.list + ";\n";
+                decl.text += modifiers2 + type.annotations + context.shorten(desugarVarargs(type.javaName)) + " " + dcl.javaName + dcl.parameters.list + ";\n";
             }
             decl.signature = dcl.signature;